example: replace logo track with screen-sharing
