+++ examples/wgpu_room/src/main.rs
@@ -4,7 +4,7 @@ use std::thread;
 use std::time::Duration;
 
 mod app;
-mod logo_track;
+mod screensharing_track;
 mod service;
 mod sine_track;
 mod video_grid;
